pyproject.toml
[tool.poetry]
name = "CobaltAPI"
version = "2.0"
description = "General purpose API with many endpoints - in active development"
authors = ["cobaltgit <criterion@chitco.co.uk>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.75.0"}
Pillow = "^9.0.1"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
autoflake = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"